requirements.txt
# Core Dependencies
torch==2.0.1
torchvision==0.15.2
torchaudio==0.15.2

# Transformers & NLP
transformers==4.36.2
datasets==2.14.5
tokenizers==0.14.1
sentencepiece==0.1.99

# Vietnamese Language Processing
vietai-vit5==1.0.0
phobert==1.0.0
pyvi==0.1.1

# Web Automation
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
selenium==4.15.1  # Fallback

# Vector Databases
faiss-cpu==1.7.4  # Use faiss-gpu for GPU
chromadb==0.4.14
pinecone-client==2.2.4  # Optional for cloud

# Machine Learning & Optimization
peft==0.7.0  # LoRA fine-tuning
bitsandbytes==0.41.1
accelerate==0.24.1
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1
tqdm==4.66.1

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Async & Concurrency
aiohttp==3.9.1
asyncio==3.4.3
anyio==4.0.0

# Data Processing
pandas==2.1.3
polars==0.19.11

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
imageio==2.33.1
ultralytics==8.1.0

# Configuration Management
omegaconf==2.3.0
hydra-core==1.3.4

# Development Dependencies (optional)
# Install with: pip install -r requirements-dev.txt

pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.12.0
isort==5.13.2
pylint==3.0.3
mypy==1.7.1
flake8==6.1.0
pre-commit==3.5.0
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
jupyter==1.0.0
ipython==8.18.1
notebook==7.0.6

# Version Control
gitpython==3.1.40

# API & Web Framework (optional for deployment)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database (optional)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Deployment
gunicorn==21.2.0
docker==7.0.0

# Profiling & Performance Analysis
memory-profiler==0.61.0
line-profiler==4.1.1

# Cloud Integration (optional)
boto3==1.29.7  # AWS
google-cloud-storage==2.10.0  # GCP
azure-storage-blob==12.19.0  # Azure

# Notes:
# - torch: Install with CUDA support: pip install torch --index-url https://download.pytorch.org/whl/cu118
# - faiss-cpu: Use faiss-gpu==1.7.4 for GPU acceleration
# - phobert & vit5: Download during setup.py or manually
# - playwright: Run "playwright install chromium" after pip install
